Update protocol to support other bulb types
The current implementation only supports RGB+W bulbs, but the protocol
supports multiple other types. Add support for identifying the controller
based on its ID byte, and add a real rgbw command for devices that handle
it. In addition, there's some amount of support for devices that handle
warm+cold white - this is entirely untested and unlikely to work well.